--- common/custom_tools.py
@@ -1,104 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the terms described in the LICENSE file in
-# the root directory of this source tree.
-
-import json
-
-from abc import abstractmethod
-from typing import Dict, List, Union
-
-from llama_stack_client.types import *  # noqa: F403
-from llama_stack_client.types.agent_create_params import *  # noqa: F403
-from llama_stack_client.types.tool_param_definition_param import *  # noqa: F403
-from typing_extensions import TypeAlias
-
-Message: TypeAlias = Union[UserMessage, ToolResponseMessage]
-
-
-class CustomTool:
-    """
-    Developers can define their custom tools that models can use
-    by extending this class.
-
-    Developers need to provide
-        - name
-        - description
-        - params_definition
-        - implement tool's behavior in `run_impl` method
-
-    NOTE: The return of the `run` method needs to be json serializable
-    """
-
-    @abstractmethod
-    def get_name(self) -> str:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_description(self) -> str:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_params_definition(self) -> Dict[str, ToolParamDefinitionParam]:
-        raise NotImplementedError
-
-    def get_instruction_string(self) -> str:
-        return f"Use the function '{self.get_name()}' to: {self.get_description()}"
-
-    def parameters_for_system_prompt(self) -> str:
-        return json.dumps(
-            {
-                "name": self.get_name(),
-                "description": self.get_description(),
-                "parameters": {
-                    name: definition.__dict__
-                    for name, definition in self.get_params_definition().items()
-                },
-            }
-        )
-
-    def get_tool_definition(self) -> AgentConfigToolFunctionCallToolDefinition:
-        return AgentConfigToolFunctionCallToolDefinition(
-            function_name=self.get_name(),
-            description=self.get_description(),
-            parameters=self.get_params_definition(),
-            type="function_call",
-        )
-
-    @abstractmethod
-    async def run(self, messages: List[Message]) -> List[Message]:
-        raise NotImplementedError
-
-
-class SingleMessageCustomTool(CustomTool):
-    """
-    Helper class to handle custom tools that take a single message
-    Extending this class and implementing the `run_impl` method will
-    allow for the tool be called by the model and the necessary plumbing.
-    """
-
-    async def run(self, messages: List[CompletionMessage]) -> List[ToolResponseMessage]:
-        assert len(messages) == 1, "Expected single message"
-
-        message = messages[0]
-
-        tool_call = message.tool_calls[0]
-
-        try:
-            response = await self.run_impl(**tool_call.arguments)
-            response_str = json.dumps(response, ensure_ascii=False)
-        except Exception as e:
-            response_str = f"Error when running tool: {e}"
-
-        message = ToolResponseMessage(
-            call_id=tool_call.call_id,
-            tool_name=tool_call.tool_name,
-            content=response_str,
-            role="ipython",
-        )
-        return [message]
-
-    @abstractmethod
-    def run_impl(self, *args, **kwargs):
-        raise NotImplementedError()
