+++ examples/custom_tools/ticker_data.py
@@ -3,49 +3,14 @@
 #
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
-import json
-from abc import abstractmethod
-from typing import Dict, List
+from typing import Dict
 
 import yfinance as yf
-from llama_stack_client.lib.agents.custom_tool import CustomTool
-from llama_stack_client.types import CompletionMessage, ToolResponseMessage
 from llama_stack_client.types.tool_param_definition_param import (
     ToolParamDefinitionParam,
 )
 
-
-class SingleMessageCustomTool(CustomTool):
-    """
-    Helper class to handle custom tools that take a single message
-    Extending this class and implementing the `run_impl` method will
-    allow for the tool be called by the model and the necessary plumbing.
-    """
-
-    async def run(self, messages: List[CompletionMessage]) -> List[ToolResponseMessage]:
-        assert len(messages) == 1, "Expected single message"
-
-        message = messages[0]
-
-        tool_call = message.tool_calls[0]
-
-        try:
-            response = await self.run_impl(**tool_call.arguments)
-            response_str = json.dumps(response, ensure_ascii=False)
-        except Exception as e:
-            response_str = f"Error when running tool: {e}"
-
-        message = ToolResponseMessage(
-            call_id=tool_call.call_id,
-            tool_name=tool_call.tool_name,
-            content=response_str,
-            role="ipython",
-        )
-        return [message]
-
-    @abstractmethod
-    async def run_impl(self, *args, **kwargs):
-        raise NotImplementedError()
+from .single_message import SingleMessageCustomTool
 
 
 class TickerDataTool(SingleMessageCustomTool):
