+++ tests/example_custom_tool.py
@@ -3,26 +3,43 @@
 #
 # This source code is licensed under the terms described in the LICENSE file in
 # the root directory of this source tree.
+import json
+from typing import Dict, List
 
-import os
-import sys
-
-THIS_DIR = os.path.dirname(__file__)
-sys.path += os.path.abspath(THIS_DIR + "../")
-
-from typing import Dict
-
-from common.custom_tools import SingleMessageCustomTool
-
-from llama_models.llama3.api.datatypes import ToolParamDefinition
+from llama_stack_client.lib.agents.custom_tool import CustomTool
+from llama_stack_client.types import CompletionMessage, ToolResponseMessage
+from llama_stack_client.types.tool_param_definition_param import (
+    ToolParamDefinitionParam,
+)
 
 
-class GetBoilingPointTool(SingleMessageCustomTool):
+class GetBoilingPointTool(CustomTool):
     """Tool to give boiling point of a liquid
     Returns the correct value for water in Celcius and Fahrenheit
     and returns -1 for other liquids
 
     """
+
+    async def run(self, messages: List[CompletionMessage]) -> List[ToolResponseMessage]:
+        assert len(messages) == 1, "Expected single message"
+
+        message = messages[0]
+
+        tool_call = message.tool_calls[0]
+
+        try:
+            response = await self.run_impl(**tool_call.arguments)
+            response_str = json.dumps(response, ensure_ascii=False)
+        except Exception as e:
+            response_str = f"Error when running tool: {e}"
+
+        message = ToolResponseMessage(
+            call_id=tool_call.call_id,
+            tool_name=tool_call.tool_name,
+            content=response_str,
+            role="ipython",
+        )
+        return [message]
 
     def get_name(self) -> str:
         return "get_boiling_point"
@@ -30,12 +47,12 @@ class GetBoilingPointTool(SingleMessageCustomTool):
     def get_description(self) -> str:
         return "Get the boiling point of a imaginary liquids (eg. polyjuice)"
 
-    def get_params_definition(self) -> Dict[str, ToolParamDefinition]:
+    def get_params_definition(self) -> Dict[str, ToolParamDefinitionParam]:
         return {
-            "liquid_name": ToolParamDefinition(
+            "liquid_name": ToolParamDefinitionParam(
                 param_type="string", description="The name of the liquid", required=True
             ),
-            "celcius": ToolParamDefinition(
+            "celcius": ToolParamDefinitionParam(
                 param_type="boolean",
                 description="Whether to return the boiling point in Celcius",
                 required=False,
